feat: Add interpolated pose calculation function
Add a new function `calcInterpolatedPoseWithVelocity` to calculate an interpolated pose with velocity for a given relative time along a path. This function takes a vector of `PoseWithVelocityStamped` objects representing the path and the relative time as input. It returns an optional `PoseWithVelocityStamped` object containing the interpolated pose, velocity, and time if the interpolation is successful. Otherwise, it returns `std::nullopt`.

Co-authored-by: kyoichi-sugahara <kyoichi.sugahara@tier4.jp>
Signed-off-by: kyoichi-sugahara <kyoichi.sugahara@tier4.jp>