refactor: update updateOdom function and test files
This commit updates the updateOdom function in the test_controller_node.cpp file and the trajectory_follower_test_utils.hpp file. It fixes a typo in the function and improves the generation of clothoid trajectories. Additionally, it updates the wait time in the test file. These changes were made to enhance the functionality and accuracy of the code.

Signed-off-by: kyoichi-sugahara <kyoichi.sugahara@tier4.jp>