chore: Update publish_debug_data function in MPC class
Update the publish_debug_data function in the MPC class to include additional debug data for debugging purposes. This change adds new variables and updates the existing ones in the MpcDebug message. The opt_error variable is now split into opt_error_updated and opt_error_before_update, and the corresponding Float64MultiArrayStamped variables are also updated. The number of time steps, num_step, is now passed as a parameter to the function.

Signed-off-by: kyoichi-sugahara <kyoichi.sugahara@tier4.jp>